Reorder directories. Rework config_test.cpp to use library
