+++ src/libtest/libtest.cpp
@@ -11,11 +11,15 @@
 #include <stdio.h>
 #include <assert.h>
 
-#include "libtest.hpp"
+#include "libtest/libtest.hpp"
 
-// TODO use the logger. Make logger a library?
+// TODO use the logger.
+// testEquals
+// specify log level in command line
+// pretty printing
+// directory issues
 
-static test_stats global_stats;
+static test_stats global_stats = {0};
 static test_stats func_stats;
 
 static const char *test_suite_name;
@@ -23,13 +27,17 @@ static const char *test_suite_name;
 int testlib_init(const char *suite_name) {
 	test_suite_name = suite_name;
 
-	printf("BEGIN SUITE %32s\n", test_suite_name);
-	printf("Log file is %32s\n", "TODO");
+	printf("BEGIN SUITE %.32s\n", test_suite_name);
+	printf("Log file is %.32s\n", "TODO");
 
 	return (0);
 }
 
 int testlib_shutdown() {
+	printf("END SUITE %.32s. pass %d, fail %d, total %d\n",
+			test_suite_name,global_stats.num_pass,
+			global_stats.num_fail, global_stats.num_total);
+
 	return (0);
 }
 
@@ -37,21 +45,27 @@ int test(const char *test_name, int (*test_func)(void *, TestStats), void *test_
 	int result;
 	test_stats stats = {0};
 
-	printf("BEGIN TEST %32s: %32s\n", test_suite_name, test_name);
+	printf("BEGIN TEST %.32s: %.32s\n", test_suite_name, test_name);
 	result = test_func(test_args, &stats);
 
-	printf("END TEST %32s: %32s. pass %d, fail %d, total %d\n",\
-			test_suite_name, test_name,\
+	printf("END TEST %.32s: %.32s. pass %d, fail %d, total %d\n",
+			test_suite_name, test_name,
 			stats.num_pass, stats.num_fail, stats.num_total);
+
+
+	global_stats.num_pass += stats.num_pass;
+	global_stats.num_fail += stats.num_fail;
+	global_stats.num_total += stats.num_total;
+
 	return (result);
 }
 
 int assert_true(int result, TestStats s, const char *assert_name) {
 	if (result) {
-		printf("\t %32s pass\n", assert_name);
+		printf("\t %.32s pass\n", assert_name);
 		s->num_pass++;
 	} else {
-		printf("\t %32s fail\n", assert_name);
+		printf("\t %.32s fail\n", assert_name);
 		s->num_fail++;
 	}
