+++ test/config/CMakeLists.txt
@@ -1,0 +1,3 @@
+# Create the config test executables
+add_executable(config_test config_test.cpp)
+target_link_libraries(config_test libtest config)
